+++ Process/Instruction/Sub.cpp
@@ -1,6 +1,6 @@
 #include "Sub.h"
 
-Sub::Sub(int var, int var1, int var2) : Instruction(SUB){
+Sub::Sub(uint16_t var, uint16_t var1, uint16_t var2) : Instruction(SUB){
     this->var = var;
     this->var1 = var1;
     this->var2 = var2;
